refactor follower, frind, friend request, add block feature and add policies

+++ SocialMedia.Api/Controllers/FriendController.cs
@@ -6,6 +6,7 @@ using SocialMedia.Data.DTOs;
 using SocialMedia.Data.Models;
 using SocialMedia.Data.Models.ApiResponseModel;
 using SocialMedia.Data.Models.Authentication;
+using SocialMedia.Repository.BlockRepository;
 using SocialMedia.Service.FriendsService;
 
 namespace SocialMedia.Api.Controllers
@@ -17,11 +18,15 @@ namespace SocialMedia.Api.Controllers
 
         private readonly IFriendService _friendService;
         private readonly UserManager<SiteUser> _userManager;
-        public FriendController(IFriendService _friendService, UserManager<SiteUser> _userManager)
+        private readonly IBlockRepository _blockRepository;
+        public FriendController(IFriendService _friendService, UserManager<SiteUser> _userManager,
+            IBlockRepository _blockRepository)
         {
             this._friendService = _friendService;
             this._userManager = _userManager;
+            this._blockRepository = _blockRepository;
         }
+
 
         [HttpGet("friends/{userName}")]
         public async Task<IActionResult> GetFriensdByUserNameAsync([FromRoute] string userName)
@@ -38,7 +43,7 @@ namespace SocialMedia.Api.Controllers
                         if(await _userManager.IsInRoleAsync(user, "Admin")
                             || user.Id == userByUserName.Id || !userByUserName.IsFriendListPrivate)
                         {
-                            var response = await _friendService.GetAllUserFriendsAsync(user.Id);
+                            var response = await _friendService.GetAllUserFriendsAsync(userByUserName.Id);
                             return Ok(response);
                         }
                     }
@@ -106,7 +111,7 @@ namespace SocialMedia.Api.Controllers
             }
         }
 
-        [HttpDelete("deleteFriend/{friendIdOrUserNameOrEmail}")]
+        [HttpDelete("unFriend/{friendIdOrUserNameOrEmail}")]
         public async Task<IActionResult> DeleteFriendAsync([FromRoute] string friendIdOrUserNameOrEmail)
         {
             {
@@ -119,14 +124,28 @@ namespace SocialMedia.Api.Controllers
                         var routeUser = await GetUserAsync(friendIdOrUserNameOrEmail);
                         if (user != null && routeUser != null)
                         {
-                            var response = await _friendService.DeleteFriendAsync(user.Id, routeUser.Id);
-                            return Ok(response);
+                            if (user.Id != routeUser.Id)
+                            {
+                                var isBlocked = await _blockRepository.GetBlockByUserIdAndBlockedUserIdAsync(
+                                user.Id, routeUser.Id);
+                                if (isBlocked == null)
+                                {
+                                    var response = await _friendService.DeleteFriendAsync(user.Id, routeUser.Id);
+                                    return Ok(response);
+                                }
+                            }
+                            return StatusCode(StatusCodes.Status403Forbidden, new ApiResponse<string>
+                            {
+                                StatusCode = 403,
+                                IsSuccess = false,
+                                Message = "Forbidden"
+                            });
                         }
-                        return StatusCode(StatusCodes.Status403Forbidden, new ApiResponse<string>
+                        return StatusCode(StatusCodes.Status406NotAcceptable, new ApiResponse<string>
                         {
-                            StatusCode = 403,
-                            IsSuccess = false,
-                            Message = "Forbidden"
+                            StatusCode = 406,
+                            Message = "Not Acceptable",
+                            IsSuccess = false
                         });
                     }
                     return StatusCode(StatusCodes.Status401Unauthorized, new ApiResponse<string>
